Fix problem with recent Starcounter XSON
Recent versions of Starcounter XSON does not allow `Add` operations in JSON Patch sent from client side.

When `<juicy-select>` is dynamically inserted in a DOM based on view model structure, and the structure changes rapidly, the view model may change so that the element may be ripped out of the DOM before the `setTimeout()` on line 47 times out. When `value` is set on line 55, the view model property bound to the `value` attribute no longer exists in the view model, which triggers a JSON Patch to be sent to the server with an `Add` operation. Fail!

The proposed change protects the code in `setTimeout()` by ensuring that he element is still attached before doing anything potentially bad.